+++ app/components/IntroCard.tsx
