+++ app/types/product.ts
